feat: store match result after match end
