+++ backend/config/settings.py
@@ -39,7 +39,6 @@ INSTALLED_APPS = [
     'oauth.apps.OauthConfig',
 	'daphne',
 	'websocket',
-	'game.apps.GameConfig',
     'users.apps.UsersConfig',
     'games.apps.GamesConfig',
     'friends.apps.FriendsConfig',
